apps/gfc/requirements.txt
django-socialregistration
